parser/tests/tree_construction/basic.dat

#data
<?COM--MENT?>
#errors
(1,1): expected-tag-name-but-got-question-mark
(1,13): expected-doctype-but-got-eof
#new-errors
(1:2) unexpected-question-mark-instead-of-tag-name
#document
| <!-- ?COM--MENT? -->
| <html>
|   <head>
|   <body>

#data
<!COM--MENT>
#errors
(1,2): expected-dashes-or-doctype
(1,12): expected-doctype-but-got-eof
#new-errors
(1:3) incorrectly-opened-comment
#document
| <!-- COM--MENT -->
| <html>
|   <head>
|   <body>

#data
</ COM--MENT >
#errors
(1,2): expected-closing-tag-but-got-char
(1,14): expected-doctype-but-got-eof
#new-errors
(1:3) invalid-first-character-of-tag-name
#document
| <!--  COM--MENT  -->
| <html>
|   <head>
|   <body>
